+++ test/web/core.js
@@ -148,3 +148,141 @@ wait(function () { return done; });
 success
 {body: "", headers: {"content-type": "text/html"}, reason: "ok", status: 200}
 */
+
+// == SECTION core - proxy requests
+
+// successful proxy to local requests
+
+done = false;
+startTime = Date.now();
+var res = local.web.dispatch({ method:'get', url:'proxy:httpl://test.com/proxy|httpl://test.com' });
+res.then(printSuccess, printError).always(finishTest);
+wait(function () { return done; });
+
+/* =>
+success
+{
+  body: "service resource",
+  headers: {
+    "content-type": "text/plain",
+    link: [
+      {href: "/", rel: "self current"},
+      {href: "/events", id: "events", rel: "collection"},
+      {href: "/foo", id: "foo", rel: "collection"},
+      {href: "/{id}", rel: "collection"}
+    ],
+    via: ["httpl://test.com"]
+  },
+  reason: "ok",
+  status: 200
+}
+*/
+
+// unsuccessful proxy to local requests
+
+done = false;
+startTime = Date.now();
+var res = local.web.dispatch({ method:'get', url:'proxy:httpl://test.com/proxy|httpl://test.com/bad/url' });
+res.then(printSuccess, printError).always(finishTest);
+wait(function () { return done; });
+
+/* =>
+error
+{
+  body: "",
+  headers: {"content-type": "text/plain"},
+  reason: "bad gateway",
+  status: 502
+}
+*/
+
+// successful proxy to remote requests
+
+done = false;
+startTime = Date.now();
+var res = local.web.dispatch({ method:'get', url:'proxy:httpl://test.com/proxy|http://grimwire.com:8080', headers: { accept: 'application/json' } });
+res.then(printSuccess, printError).always(finishTest);
+wait(function () { return done; });
+
+/* =>
+success
+{
+  body: {hello: "world"},
+  headers: {
+    allow: "OPTIONS, HEAD, GET",
+    "content-type": "application/json",
+    link: [
+      {href: "/", rel: "self current"},
+      {href: "/foo", id: "foo", rel: "collection"},
+      {href: "/{id}", rel: "collection"}
+    ],
+    via: ["httpl://test.com"]
+  },
+  reason: "Ok",
+  status: 200
+}
+*/
+
+// unsuccessful proxy to remote requests
+
+done = false;
+startTime = Date.now();
+var res = local.web.dispatch({ method:'get', url:'proxy:httpl://test.com/proxy|http://grimwire.com:8080/bad/url' });
+res.then(printSuccess, printError).always(finishTest);
+wait(function () { return done; });
+
+/* =>
+error
+{
+  body: "",
+  headers: {"content-type": "text/plain"},
+  reason: "bad gateway",
+  status: 502
+}
+*/
+
+// successful two-stage proxy to local requests
+
+done = false;
+startTime = Date.now();
+var res = local.web.dispatch({ method:'get', url:'proxy:httpl://test.com/proxy|httpl://test.com/proxy|httpl://test.com' });
+res.then(printSuccess, printError).always(finishTest);
+wait(function () { return done; });
+
+/* =>
+success
+{
+  body: "service resource",
+  headers: {
+    "content-type": "text/plain",
+    link: [
+      {href: "/", rel: "self current"},
+      {href: "/events", id: "events", rel: "collection"},
+      {href: "/foo", id: "foo", rel: "collection"},
+      {href: "/{id}", rel: "collection"}
+    ],
+    via: ["httpl://test.com"]
+  },
+  reason: "ok",
+  status: 200
+}
+*/
+
+
+// unsuccessful two-stage proxy to local requests
+
+done = false;
+startTime = Date.now();
+var res = local.web.dispatch({ method:'get', url:'proxy:httpl://test.com/proxy|httpl://test.com/proxy|httpl://test.com/bad/url' });
+res.then(printSuccess, printError).always(finishTest);
+wait(function () { return done; });
+
+/* =>
+error
+{
+  body: "",
+  headers: {"content-type": "text/plain"},
+  reason: "bad gateway",
+  status: 502
+}
+*/